backend/requirements.txt
django>=5.0,<6.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
